Updated symlinks script.

+++ make-symlinks.py
@@ -2,29 +2,45 @@
 
 import os
 
-def make_symlink(git_filename, os_filepath):
+def make_symlink(git_filename, os_filepath, sudo=False, soft=False):
     home = os.path.expanduser('~/')
     cwd = os.getcwd()
     relpath = os.path.relpath(cwd, home)
 
     git_config = home+relpath+'/'+git_filename
     config = os.path.expanduser(os_filepath)
-    try:
-        os.remove(config)
-    except OSError:
-        pass
-    os.symlink(git_config, config)
+
+    if os.path.isfile(config):
+        if sudo:
+            os.system('sudo rm %s' % config)
+        else:
+            os.remove(config)
+
+    if sudo and not soft:
+        os.system('sudo ln %s %s' % (git_config, config))
+    elif sudo and soft:
+        os.system('sudo ln -s %s %s' % (git_config, config))
+    elif not sudo and not soft:
+        os.system('ln %s %s' % (git_config, config))
+    elif not sudo and soft:
+        os.system('ln -s %s %s' % (git_config, config))
 
 def main():
     termite_config = os.path.expanduser('~/.config/termite/config')
+    i3_config = os.path.expanduser('~/.config/i3/config')
+    polybar_config = os.path.expanduser('~/.config/polybar/config')
+    rofi_config = os.path.expanduser('~/.config/rofi/config')
     nvim_config = os.path.expanduser('~/.nvimrc')
     zshrc_config = os.path.expanduser('~/.zshrc')
-    desktop_files = '/usr/share/applications'
+    desktop_files = '/usr/share/applications/nvim-term.desktop'
 
-    make_symlink('termite-config', termite_config)
-    make_symlink('vimrc', nvim_config)
-    make_symlink('zshrc', zshrc_config)
-    make_symlink('nvim-term.desktop', desktop_files)
+    make_symlink('termite-config', termite_config, soft=True)
+    make_symlink('i3-config', i3_config, soft=True)
+    make_symlink('polybar-config', polybar_config, soft=True)
+    make_symlink('rofi-config', rofi_config, soft=True)
+    make_symlink('nvimrc', nvim_config, soft=True)
+    make_symlink('zshrc', zshrc_config, soft=True)
+    make_symlink('nvim-term.desktop', desktop_files, sudo=True, soft=True)
 
 if __name__ == '__main__':
     main()
